+++ Board_prototype/Assets/Scripts/UI/Text/TextMover.cs
@@ -7,7 +7,7 @@ public class TextMover : MonoBehaviour
     private bool isActive = false;
     Vector2 direction;
 
-    public void setup(int scoreValue)
+    public void setup(System.UInt64 scoreValue)
     {
         GetComponent<Text>().text = "+" + scoreValue;
         direction = Vector2.up * 150;
